+++ packages/@expo/cli/src/customize/typescript.ts
@@ -20,7 +20,7 @@ export async function typescript(projectRoot: string) {
   }
 
   const { exp } = getConfig(projectRoot, { skipSDKVersionRequirement: true });
-  await new MetroBundlerDevServer(projectRoot, getPlatformBundlers(exp), {
+  await new MetroBundlerDevServer(projectRoot, getPlatformBundlers(projectRoot, exp), {
     isDevClient: true,
   }).startTypeScriptServices();
 }