+++ packages/@expo/cli/src/start/server/webpack/__tests__/WebpackBundlerDevServer-test.ts
@@ -23,7 +23,10 @@ afterAll(() => {
 });
 
 async function getStartedDevServer(options: Partial<BundlerStartOptions> = {}) {
-  const devServer = new WebpackBundlerDevServer('/', getPlatformBundlers({}));
+  const devServer = new WebpackBundlerDevServer(
+    '/',
+    getPlatformBundlers('/', { web: { bundler: 'webpack' } })
+  );
   devServer['getAvailablePortAsync'] = jest.fn(() => Promise.resolve(3000));
   // Tested in the superclass
   devServer['postStartAsync'] = jest.fn(async () => {});
@@ -33,7 +36,10 @@ async function getStartedDevServer(options: Partial<BundlerStartOptions> = {}) {
 
 describe('bundleAsync', () => {
   it(`bundles in dev mode`, async () => {
-    const devServer = new WebpackBundlerDevServer('/', getPlatformBundlers({}));
+    const devServer = new WebpackBundlerDevServer(
+      '/',
+      getPlatformBundlers('/', { web: { bundler: 'webpack' } })
+    );
 
     devServer['clearWebProjectCacheAsync'] = jest.fn();
     devServer['loadConfigAsync'] = jest.fn(async () => ({}));
@@ -89,12 +95,18 @@ describe('startAsync', () => {
 describe('getProjectConfigFilePath', () => {
   it(`loads from project`, async () => {
     vol.fromJSON({ 'webpack.config.js': '{}' }, '/');
-    const devServer = new WebpackBundlerDevServer('/', getPlatformBundlers({}));
+    const devServer = new WebpackBundlerDevServer(
+      '/',
+      getPlatformBundlers('/', { web: { bundler: 'webpack' } })
+    );
     expect(devServer.getProjectConfigFilePath()).toBe('/webpack.config.js');
   });
   it(`cannot load from project`, async () => {
     vol.fromJSON({ 'package.json': '{}' }, '/');
-    const devServer = new WebpackBundlerDevServer('/', getPlatformBundlers({}));
+    const devServer = new WebpackBundlerDevServer(
+      '/',
+      getPlatformBundlers('/', { web: { bundler: 'webpack' } })
+    );
     expect(devServer.getProjectConfigFilePath()).toBe(null);
   });
 });